A "mostly working" UM source and sink connector
Debugged the UMSourceTask and added an LBM source to the UM SinkTask. It mostly works, i.e., I can send in LBM messages to the UM source connector on a topic "um-in" and receive LBM messages from an LBM sink connector on "um-out".  Unfortunately the Kafka record has a '2' prepended to messages.